refactor: barbican: diverse the functions logic
